requirements.txt
albumentations==1.0.3
apturl==0.5.2
asgiref==3.4.1
attrs==19.3.0
backcall==0.2.0
bcrypt==3.1.7
bleach==3.1.1
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
debugpy==1.4.1
decorator==5.0.9
defer==1.0.6
defusedxml==0.6.0
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
future==0.18.2
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
imageio==2.9.0
importlib-metadata==1.5.0
ipykernel==6.0.3
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==6.0.0
jedi==0.18.0
Jinja2==2.10.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
kaggle==1.5.12
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.4
networkx==2.6.2
nltk==3.6.2
notebook==6.0.3
numpy==1.21.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.3.56
opencv-python-headless==4.5.3.56
pandas==1.3.1
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.8.2
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.19
protobuf==3.6.1
pycairo==1.16.2
pycups==1.9.73
Pygments==2.9.0
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.15.5
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-slugify==5.0.2
pytz==2019.3
PyWavelets==1.1.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==22.1.0
regex==2021.8.3
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
SecretStorage==2.3.1
Send2Trash==1.5.0
simplejson==3.16.0
six==1.14.0
sklearn==0.0
sqlparse==0.4.1
systemd-python==234
tabulate==0.8.9
terminado==0.8.2
testpath==0.4.4
text-unidecode==1.3
textblob==0.15.3
threadpoolctl==2.2.0
tifffile==2021.8.8
torch==1.9.0
torchaudio==0.9.0
torchvision==0.10.0
tornado==6.1
tqdm==4.62.0
traitlets==5.0.5
typing-extensions==3.10.0.0
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==2.0.0
wordcloud==1.8.1
xkit==0.0.0
zipp==1.0.0
